requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.29
langchain-openai==0.3.11
langchain-pinecone==0.2.8
langchain-groq==0.2.1
langchain-huggingface==0.1.2
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
-e .